+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Added `compas_occ.geometry.NurbsCurve.joined`.
 
 ### Changed
+* Extend `compas_occ.geometry.OCCNurbsSurface.from_fill` with up to 4 input curves.
 
 ### Removed